From  CSV file representing a list of planet found by Kepler,
used the Node FS module to read the file as a stream of data, then, used the third-party module CSV parse to parse the CSV file and obtain an array of object representing each planet found. The filtered out this array of objects to potentially habitable planets using different criteria. (sorry for the bad english)